moved genesis block generation out of forger
