+++ node/src/main/scala/co/topl/consensus/Forger.scala
@@ -112,78 +112,6 @@ object Forger {
       Future.successful(Done)
     }
 
-  def genesisBlock(
-    settings:                    AppSettings,
-    networkType:                 NetworkType,
-    fetchStartupKeyView:         () => Future[StartupKeyView],
-    consensusVariablesInterface: ConsensusVariablesHolder
-  )(implicit
-    nxtLeaderElection: NxtLeaderElection,
-    system:            ActorSystem[_],
-    ec:                ExecutionContext
-  ): Future[Block] = {
-    implicit val networkPrefix: NetworkPrefix = networkType.netPrefix
-
-    def initializeFromChainParamsAndGetBlock(block: Try[(Block, ChainParams)]): Future[Block] = {
-
-      import scala.concurrent.duration._
-      implicit val timeout: Timeout = Timeout(10.seconds)
-
-      Future.fromTry(block).flatMap { case (block: Block, ChainParams(totalStake, initDifficulty)) =>
-        consensusVariablesInterface
-          .update(block.id, ConsensusParamsUpdate(Some(totalStake), Some(initDifficulty), Some(0L), Some(0L)))
-          .valueOrF(e => Future.failed(e.reason))
-          .map(_ => block)
-      }
-    }
-
-    if (settings.forging.genesis.head.genesisStrategy.isEmpty) {
-      networkType match {
-        case Mainnet         => initializeFromChainParamsAndGetBlock(ToplnetGenesis.getGenesisBlock)
-        case ValhallaTestnet => initializeFromChainParamsAndGetBlock(ValhallaGenesis.getGenesisBlock)
-        case HelTestnet      => initializeFromChainParamsAndGetBlock(HelGenesis.getGenesisBlock)
-        case PrivateTestnet =>
-          fetchStartupKeyView()
-            .map(view => PrivateGenesis(view.addresses, settings).getGenesisBlock)
-            .flatMap(r => initializeFromChainParamsAndGetBlock(r))
-        case _ =>
-          Future.failed(new IllegalArgumentException(s"Undefined network type $networkType"))
-      }
-    } else {
-      settings.forging.genesis.head.genesisStrategy.head match {
-        case Generated =>
-          if (networkType == PrivateTestnet) {
-            settings.forging.genesis.flatMap(_.generated) match {
-              case Some(genesisSettings) =>
-                fetchStartupKeyView()
-                  .map(view =>
-                    GeneratedGenesis(view.addresses, genesisSettings).getGenesisBlock
-                  )
-                  .flatMap(r => initializeFromChainParamsAndGetBlock(r))
-              case None =>
-                Future.failed(new IllegalArgumentException("Failed to read generated genesis settings"))
-            }
-          } else {
-            Future.failed(new IllegalArgumentException(s"Unsupported network type for generated genesis: $networkType"))
-          }
-        case FromBlockJson =>
-          settings.forging.genesis.flatMap(_.fromBlockJson) match {
-            case Some(genesisSettings) =>
-              initializeFromChainParamsAndGetBlock(
-                GenesisFromBlockJson(genesisSettings, networkType, nxtLeaderElection.protocolMngr).getGenesisBlock
-              )
-            case None =>
-              Future.failed(new IllegalArgumentException("Failed to read genesis from block Json settings"))
-          }
-
-        case _ =>
-          Future.failed(
-            new IllegalArgumentException(s"Undefined genesis strategy $settings.forging.genesis.head.genesisStrategy")
-          )
-      }
-    }
-  }
-
   sealed trait ForgerFailure
   case class ForgeFailure(forgeFailure: Forge.Failure) extends ForgerFailure
   case class ForgingError(error: Throwable) extends ForgerFailure
